requirements.txt
pytest==9.0.2
allure-pytest==2.15.3
selene>=2.0.0b1
selenium==4.41.0